requirements.txt
discord.py==2.5.2
python-dotenv==1.0.1
python-telegram-bot[webhooks]==21.11.1
